+++ beehive_service_netaas/networkservice/__init__.py
@@ -1,15 +1,28 @@
 # SPDX-License-Identifier: EUPL-1.2
 #
 # (C) Copyright 2020-2022 Regione Piemonte
+# (C) Copyright 2018-2023 CSI-Piemonte
 
-from beehive_service_netaas.networkservice.controller import ApiNetworkGateway, ApiNetworkService, \
-    ApiNetworkHealthMonitor, ApiNetworkTargetGroup, ApiNetworkListener, ApiNetworkLoadBalancer
+from beehive_service_netaas.networkservice.controller import (
+    ApiNetworkGateway,
+    ApiNetworkService,
+    ApiNetworkHealthMonitor,
+    ApiNetworkTargetGroup,
+    ApiNetworkListener,
+    ApiNetworkLoadBalancer,
+    ApiSshGateway,
+)
 from beehive_service_netaas.networkservice.views import NetworkServiceAPI
 from beehive_service_netaas.networkservice.views.gateway import NetworkGatewayAPI
-from beehive_service_netaas.networkservice.views.loadbalancer import NetworkLoadBalancerAPI
-from beehive_service_netaas.networkservice.views.securitygroup import NetworkSecurityGroupAPI
+from beehive_service_netaas.networkservice.views.loadbalancer import (
+    NetworkLoadBalancerAPI,
+)
+from beehive_service_netaas.networkservice.views.securitygroup import (
+    NetworkSecurityGroupAPI,
+)
 from beehive_service_netaas.networkservice.views.subnet import NetworkSubnetAPI
 from beehive_service_netaas.networkservice.views.vpc import NetworkVpcAPI
+from beehive_service_netaas.networkservice.views.sshgateway import NetworkSshGatewayAPI
 
 
 class NetworkServicePlugin(object):
@@ -22,13 +35,14 @@ class NetworkServicePlugin(object):
             ApiNetworkTargetGroup,
             ApiNetworkListener,
             ApiNetworkLoadBalancer,
+            ApiSshGateway,
         ]
-    
+
     def init(self):
         for srv in self.st_plugins:
             service = srv(self.module.get_controller())
             service.init_object()
-    
+
     def register(self):
         apis = [
             NetworkServiceAPI,
@@ -37,5 +51,6 @@ class NetworkServicePlugin(object):
             NetworkVpcAPI,
             NetworkSubnetAPI,
             NetworkSecurityGroupAPI,
+            NetworkSshGatewayAPI,
         ]
         self.module.set_apis(apis)